Widen HII feedback cost fields from float to double
Legitimate photon-count values in this run's internal units (as low as
~9e55 from the N_HI floor alone, as high as ~6.9e59) exceed FLT_MAX
(~3.4e38), guaranteeing overflow to +inf on the narrowing cast under
MPI. Cross-checked via CGS conversion and Hui and Gnedin 1997: the
magnitude itself is physically sane, this is a pure wire-type-sizing
bug, not a units or physics error.

feedback_part_data.cost and hii_tag_report.cost move to double, with
matching signature updates across every feedback flavor's
feedback_unpack_hii_tag_report stub. Field order in both structs
places cost next to an existing double-aligned field, so struct size
is unchanged (48/72 bytes) rather than growing from alignment
padding.

+++ src/cell_pack.c
@@ -364,7 +364,7 @@ void cell_unpack_part_hii_tag(struct cell *c, const struct hii_tag_report *data,
 
   for (size_t i = 0; i < count; ++i) {
     char collision = 0;
-    float forfeited_cost = 0.f;
+    double forfeited_cost = 0.0;
     feedback_unpack_hii_tag_report(&parts[i], &data[i], &collision,
                                    &forfeited_cost);
     if (collision) {
